stylefn: cast a bitwise not, squashing a -Wconversion warning

+++ lib/common/htmllex.c
@@ -175,7 +175,7 @@ static int stylefn(htmldata_t * p, char *v)
     for (tk = strtok (buf, DELIM); tk; tk = strtok (NULL, DELIM)) {
 	if (!strcasecmp(tk, "ROUNDED")) p->style |= ROUNDED;
 	else if (!strcasecmp(tk, "RADIAL")) p->style |= RADIAL;
-	else if(!strcasecmp(tk,"SOLID")) p->style &= ~(DOTTED|DASHED);
+	else if(!strcasecmp(tk,"SOLID")) p->style &= (unsigned short)~(DOTTED|DASHED);
 	else if(!strcasecmp(tk,"INVISIBLE") || !strcasecmp(tk,"INVIS")) p->style |= INVISIBLE;
 	else if(!strcasecmp(tk,"DOTTED")) p->style |= DOTTED;
 	else if(!strcasecmp(tk,"DASHED")) p->style |= DASHED;
